Only save the notes if the contents are not empty
This fixes a very severe bug which would cause the notes to be
saved as empty notes. I've lost all my notes from the last 10 days
because of this stupid bug.

The QTextDocument would emit the modified signal, on calling reset(),
which would result in the note being re-saved, but since the contents
are now empty, it would be saved as an emtpy note.