Update the sidebar if a note is modified or created
Ideally, we should just use the ResourceWatcher instead of the
notewidget informing us of the changes, as some other application
could also be creating notes. But since nepomuklib doesn't
currently work with Nepomuk::Resource, we'll have to stick with
this method.

I'm using a QSortFilterProxyModel to keep the list of notes sorted.
It seems a bit of an overkil. Maybe I could use some kind of
priority queue?