Fix the new note screwing up the model index bug
This is/was a super annoying bug that would cause the wrong index,
to be given after a new note was added. All the model indexes were
off by 1.

I still haven't figured out why this was happening, but I've hacked
around it. This fix is horribly in-efficient and recreates the entire
model each time a new note is added.